Renamed panel-group to tracked-panels
