+++ src/utils/tracked-panels.js
